Add new chat message filter for chat emojis in the frontend
