+++ src/typescript/sdk/src/emoji_data/index.ts
@@ -1,3 +1,4 @@
+export * from "./chat-message";
 export * from "./emoji-data";
 export * from "./types";
 export * from "./utils";
